+++ hacker-stories/src/App.jsx
@@ -1,43 +1,13 @@
-const list = [
-  {
-    title: "React",
-    url: "https://reactjs.org",
-    author: "Jordan Walker",
-    num_comments: 3,
-    points: 4,
-    objectId: 0,
-  },
-  {
-    title: "Redux",
-    url: "https://redux.org",
-    author: "Dan Abramov, Andrew Clark",
-    num_comments: 2,
-    points: 5,
-    objectId: 1,
-  },
-];
+import List from "./components/List";
+import Search from "./components/Search";
 
 function App() {
   return (
     <div>
       <h1>My Hacker Stories</h1>
-      <label htmlFor="search">Search: </label>
-      <input id="search" type="text" placeholder="Search" />
+      <Search />
       <hr />
-      <ul>
-        {list.map((item) => {
-          return (
-            <li key={item.objectId}>
-              <span>
-                <a href={item.url}>{item.title}</a>
-              </span>
-              <span>{item.author}</span>
-              <span>{item.num_comments}</span>
-              <span>{item.points}</span>
-            </li>
-          );
-        })}
-      </ul>
+      <List />
     </div>
   );
 }
