+++ hacker-stories/src/components/Search.jsx
@@ -1,0 +1,10 @@
+function Search() {
+  return (
+    <div>
+      <label htmlFor="search">Search: </label>
+      <input id="search" type="text" placeholder="Search" />
+    </div>
+  );
+}
+
+export default Search;
