+++ hacker-stories/src/data.js
@@ -1,0 +1,20 @@
+const list = [
+  {
+    title: "React",
+    url: "https://reactjs.org",
+    author: "Jordan Walker",
+    num_comments: 3,
+    points: 4,
+    objectId: 0,
+  },
+  {
+    title: "Redux",
+    url: "https://redux.org",
+    author: "Dan Abramov, Andrew Clark",
+    num_comments: 2,
+    points: 5,
+    objectId: 1,
+  },
+];
+
+export default list;
